[Refactor] Satisfy clang-tidy

+++ src/common/FullJitter.cpp
@@ -1,5 +1,8 @@
 #include "FullJitter.hpp"
 #include <spdlog/spdlog.h>
+#include <chrono>
+#include <stdexcept>
+#include <random>
 
 namespace zdb {
 
@@ -10,7 +13,7 @@ std::chrono::microseconds FullJitter::jitter(std::chrono::microseconds v) {
         spdlog::error("FullJitter: Negative duration is not supported. Throwing invalid_argument.");
         throw std::invalid_argument("Negative duration is not supported");
     }
-    std::uniform_int_distribution<int> dist(0, v.count());
+    std::uniform_int_distribution<long> dist(0, v.count());
     return std::chrono::microseconds(dist(rng));
 }
